Add command line task reference
This is a **reference** -type document intended to document the general
command line syntax and arguments of command line tasks.

The advantage of a general reference page like this is that it becomes
convenient to identify and link to these arguments when discussed in
other pages.

The page incorporates and iterates on content in the command line help
and pipe_base's main.dox.

I anticipate having additional how-to-style topics to cover usage
patterns around choosing data repositories/reruns, dataId syntax,
logging, clobbering, etc.  This topic should link to those topics when
they become available.